requirements.txt
fastapi
uvicorn
pydantic
puthon-dotenv
openai